Move search bar out of the app bar
This is not relevant for every page atm.
